% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/invgamma.R
\name{dinvgamma}
\alias{dinvgamma}
\alias{qinvgamma}
\alias{pinvgamma}
\alias{rinvgamma}
\title{Inverse gamma distribution}
\usage{
dinvgamma(x, shape, scale = 1, log)

qinvgamma(p, shape, scale = 1, lower.tail = TRUE, log.p = FALSE)

pinvgamma(q, shape, scale = 1, lower.tail = TRUE, log.p = FALSE)

rinvgamma(n, shape, scale = 1)
}
\arguments{
\item{x, q}{vector of quantiles}

\item{shape, scale}{shape and scale parameters. Must be positive.}

\item{log, logp}{logical; if TRUE, probabilities/densities p are returned as log(p)}

\item{p}{vector of probabilities}

\item{lower.tail}{logical; if TRUE (default), probabilities are p[X\le X],
otherwise, P[X>x]}

\item{n}{number of observations. If length(n) > 1, the length is taken to be
the number required.}
}
\value{
dgamma gives the density, pgamma gives the distribution function,
  qgamma gives the quantile function, and rgamma generates random deviates.
}
\description{
Probability density function, cumulative distribution function, quantile
function, and random variable generator for the inverse gamma distribution
with parameters shape and scale.
}
\details{
The inverse gamma distribution with parameters shape = a and scale = b has density

       f(x) = b^a / Gamma(a) x^{-a-1} exp(-b/x)

  for x\ge 0, a>0, b>0. (Gamma(a) is the function implemented by R's gamma()
  and defined in its help.)

  The mean is \frac{b}{a-1} (if a>1) and
  the variance is \frac{b^2}{(a-1)^2(a-2)} (if a>2).

  These functions are all based on the gamma distribution.
  Please see the documentation for issues using the gamma distribution,
  e.g. issues when the shape parameter is small.
}
\examples{
dinvgamma(2,1) == dgamma(1/2,1)/2^2


}
\seealso{
\code{\link{dgamma}}
}
